Make Opaque really opaque, fix smart pointers
This removes implementations for Opaque<Arc<T>> and just adds a blanket implementation for Opaque<T>.

It also fixes Weak having the vtable or Arc, which explains the bug
I had to fix earlier in #413
