+++ facet-core/src/impls_alloc/mod.rs
@@ -1,3 +1,4 @@
+mod arc;
 mod btreemap;
-mod smartptr;
+mod rc;
 mod vec;
